+++ docs/changelog.mdx
@@ -25,6 +25,8 @@ Semantic versioning is a standardized way of versioning software projects. It us
     - Clickup: Includes actions for getting lists, creating folderless lists, getting tasks, getting task comments, and getting spaces actions
     - Wordpress: Includes actions for creating Wordpress posts and triggering new Wordpress posts
     - Github: Includes triggers for new pull requests and new issues
+    - Posthog: Includes actions for capture events / creating project
+    - Google Tasks: Includes actions for creating task
 
 * **Features:**
     - Auto Complete styling fixes
