+++ docs/pieces/apps/posthog.mdx
@@ -1,5 +1,5 @@
 ---
-title: "Discord"
+title: "Posthog"
 description: ""
 ---
 
@@ -12,7 +12,7 @@ Check https://posthog.com/docs/api/overview#personal-api-keys-recommended
 
 ## Triggers
 
-*No supported triggers yet, please let us know if you need something on Discord so we can help out* 
+*No supported triggers yet, please let us know if you need something on Posthog so we can help out* 
 
 
 ---
