--- packages/frontend/src/assets/changelog.md
@@ -1,34 +1,0 @@
-# 29-10-2022
-
-## Improvements
-- Revamped Loop debugging experience
-- Revamped collection naming
-- Add Checkbox control
-- Improved resize handle animation
-
-## Fixes
-- Fixed the editable fields to be read only while viewing runs
-
-# 20-10-2022
-
-## New Features 🚀
-
-#### Builder
-- Added Webhook Trigger
-- Added Instance Started / Stopped Trigger
-- Added Embed buttons in the builder to show exact steps.
-- Revamped publish collection experience.
-- Revamped Add Authentication Dialog.
-
-#### SDK
-- Added the ability to edit Settings after the user enables the integration.
-- Added a red star to required fields.
-- Added **Integration Disabled** event.
-
-## Improvements
-- Add Collection Storage Scope.
-- Code step show compilation errors.
-- SDK Functions wait until init function completes.
-
-## Fixes
-- Embedded UI no longer showing archived collections.
