+++ packages/pieces/apps/src/lib/csv/actions/convert-json-to-csv.ts
@@ -1,10 +1,10 @@
-import { createAction, Property, HttpRequest, HttpMethod, httpClient } from "@activepieces/framework";
+import { createAction, Property } from "@activepieces/framework";
 import { parseCSVFile } from "../utils";
 
 export const parseCSVTextAction = createAction({
-  name: 'parse_csv_text',
-  displayName: 'Parse CSV Text',
-  description: 'Read CSV and automatically parse it into a JSON array:',
+  name: "convert_csv_to_json",
+  displayName: "Convert CSV to JSON",
+  description: "This function reads a CSV string and converts it into JSON array format.",
   sampleData: [
     {
       "col1": "value",
@@ -43,13 +43,12 @@ export const parseCSVTextAction = createAction({
     }),
   },
   async run(context) {
-    const {csv_text, has_headers, delimiter_type} = context.propsValue
+    const { csv_text, has_headers, delimiter_type } = context.propsValue
     const config = {
       header: has_headers,
       delimiter: delimiter_type === "auto" ? "" : delimiter_type,
       skipEmptyLines: true
     }
-
     return parseCSVFile(csv_text, config)
   }
 });
