+++ packages/pieces/apps/src/lib/csv/index.ts
@@ -1,12 +1,11 @@
 import { createPiece } from "@activepieces/framework";
-import { parseCSVTextAction } from "./actions/parse-csv";
-import { unparseCSVTextAction } from "./actions/unparse-csv";
+import { parseCSVTextAction } from "./actions/convert-json-to-csv";
+import { unparseCSVTextAction } from "./actions/convert-csv-to-json";
 
 export const csv = createPiece({
   name: 'csv',
   displayName: "CSV",
-  //provide logo for CSV
-  logoUrl: 'https://forcetalks.s3.amazonaws.com/wp-content/uploads/2018/05/25081331/csv-logo.png',
+  logoUrl: 'https://cdn.activepieces.com/pieces/csv.png',
   version: '0.0.0',
   actions: [parseCSVTextAction, unparseCSVTextAction],
   authors: ['kanarelo'],
