+++ packages/pieces/apps/src/lib/posthog/posthog.mdx
@@ -1,5 +1,5 @@
 ---
-title: "Discord"
+title: "Posthog"
 description: ""
 ---
